probes: Introduce new networking probe
This deprecates the old location probe and replaces it by a new
networking probe.

This probe uses a (generated) file at /etc/bcfg2.netmap to detect what
sites and zones the machine is based on all its current IPv4 and IPv6
addresses.

The result is exported as groups and as probe data.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>